ref(explorer+state): refactor explorer within state module
split codes in smaller modules, export only what is necessary, use from
traits to ease node usage
